Refactor WebSocket to recv messages via mpsc
WebSocket::connect() returns a tuple with the ws, the handshake
response, an EventReceiver and a recv_loop future.

The event receiver is consumed when calling ws.enable_status_update()
which returns the initial state and a stream of future events.

The recv_loop future must run on a dedicated tokio::task before any
message can be received from the ws.